Move convexplano_lens example into module and restore integration tests
