+++ raytracer/cherry-rs/Cargo.toml
@@ -24,8 +24,3 @@ ria = { version = "0.3", optional = true }
 approx = "0.5"
 bitcode = { version = "0.6", features = ["serde"] }
 ndarray = { version = "0.16", features = [ "approx", "serde" ] }
-
-[[example]]
-name = "convexplano_lens"
-crate-type = ["staticlib"]
-test = true
