+++ raytracer/cherry-rs/src/examples/mod.rs
@@ -1,2 +1,3 @@
+//! Example lens data for various lenses and optical setups.
 pub mod convexplano_lens;
 pub mod petzval_lens;
