+++ raytracer/cherry-rs/src/views/ray_trace_3d/mod.rs
@@ -274,13 +274,18 @@ fn axial_launch_point(obj_z: Float, sur_z: Float, enp_z: Float) -> Float {
 
 #[cfg(test)]
 mod tests {
+    use crate::core::Float;
     use crate::examples::convexplano_lens::sequential_model;
+    use crate::n;
 
     use super::*;
 
     #[test]
     fn test_ray_trace_3d_view() {
-        let sequential_model = sequential_model();
+        let air = n!(1.0);
+        let nbk7 = n!(1.515);
+        let wavelengths: [Float; 1] = [0.5876];
+        let sequential_model = sequential_model(air, nbk7, &wavelengths);
 
         let aperture_spec = ApertureSpec::EntrancePupil { semi_diameter: 5.0 };
         let field_specs = vec![
